more work on jwt, still crude
